requirements.txt
Flask-Babel2
blinker
itsdangerous
log4mongo
pyTelegramBotAPI
Flask
requests
flask-restful
six
uwsgi
pymongo
mongoengine
geopy
Flask-WTF
kik
pyopenssl
pycryptodomex
python-dotenv